+++ qtisas/src/plot2D/ArrowMarker.h
@@ -140,7 +140,7 @@ public:
 private:
 	void displayInfo(bool clear = false);
     void draw(QPainter *p, const QwtScaleMap &xMap, const QwtScaleMap &yMap, const QRectF &r) const override;
-    static double theta(int xs, int ys, int xe, int ye);
+    static double theta(double xs, double ys, double xe, double ye);
 
 	//! Flag specifying if the start arrow is visible
     bool d_start_arrow{};
